+++ src/aks-preview/azext_aks_preview/_format.py
@@ -31,7 +31,7 @@ def aks_addon_list_table_format(result):
 
 
 def aks_addon_show_table_format(result):
-    def parser(entry):
+    def parser(entry): 
         config = ""
         for k, v in entry["config"].items():
             config += k + "=" + v + ";"
@@ -45,6 +45,21 @@ def aks_addon_show_table_format(result):
         return parsed.search(entry, Options(dict_cls=OrderedDict))
     return parser(result)
 
+def aks_machine_list_table_format(results):
+    return [aks_machine_show_table_format(r) for r in results]
+
+def aks_machine_show_table_format(result):
+    def parser(entry): 
+        ip_addresses = ""
+        for k in entry["properties"]["network"]["ip_addresses"].items():
+            ip_addresses += "ip:" + k["ip"] + "," + "family:" + k["family"] + ";"
+        entry["ip"] = config
+        parsed = compile_jmes("""{
+                name: name,
+                ip: ip
+            }""")
+        return parsed.search(entry, Options(dict_cls=OrderedDict))
+    return parser(result)
 
 def aks_agentpool_show_table_format(result):
     """Format an agent pool as summary results for display with "-o table"."""